requirements.txt
Flask==1.1.1
requests==2.22.0
influxdb==5.2.3
flask-cors==3.0.8
pyyaml==5.3